Openning connection viewlet after deploying database
